% Generated by roxygen2 (4.0.0): do not edit by hand
\name{addRecovery}
\alias{addRecovery}
\title{Add recovery times}
\usage{
addRecovery(g, rate)
}
\arguments{
\item{g}{an SIR graph}

\item{rate}{exponential distribution rate per unit time}
}
\value{
an SIR graph with recovery times set for infected cases
}
\description{
add recovery times
}
\details{
this function takes a graph and computes recovery times for
all infected cases based on an exponential distribution
}
\author{
Barry Rowlingson
}

