% Generated by roxygen2 (4.0.0): do not edit by hand
\name{geom_edge}
\alias{geom_edge}
\title{geom_edge}
\usage{
geom_edge(mapping = NULL, data = NULL, graph = NULL, directed = FALSE,
  ...)
}
\arguments{
\item{mapping}{the aesthetic mapping}

\item{data}{data frame}

\item{graph}{the graph}

\item{directed}{currently ignored}

\item{...}{further parameters for geom_segment}
}
\value{
a geom_segment for ggplot
}
\description{
geom for edges
}
\details{
get edge lines for a graph and plot using geom_segment
}
\author{
Barry S Rowlingson
}

