\name{geom_node}
\alias{geom_node}
\title{geom_node}
\usage{
geom_node(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, ...)
}
\arguments{
  \item{mapping}{the aesthetic mapping}

  \item{data}{the graph}

  \item{stat}{statistical transformation}

  \item{position}{position adjustment}

  \item{na.rm}{missing value behaviour}

  \item{...}{further arguments}
}
\value{
a geom for ggplot
}
\description{
geom for nodes
}
\details{
get locations from a graph and plot points. See the help
for geom_point for more.
}
\author{
Barry S Rowlingson
}

