% Generated by roxygen2 (4.0.0): do not edit by hand
\name{infectN}
\alias{infectN}
\title{infectN}
\usage{
infectN(N)
}
\arguments{
\item{N}{number of cases to create}
}
\value{
a function that resets its graph and adds N initial infections.
}
\description{
simple infection starter function
}
\details{
creates a function that creates a fixed number of cases
}
\author{
Barry S Rowlingson
}

