\name{plotSIR}
\alias{plotSIR}
\title{plotSIR}
\usage{
  plotSIR(g, layout = layout.kamada.kawai, seed = 1, ...)
}
\arguments{
  \item{g}{an SIR graph}

  \item{layout}{layout algorithm}

  \item{seed}{random seed for consistent layouts}

  \item{...}{further layout parameters}
}
\value{
  nothing
}
\description{
  SIR graph base graphics plot
}
\details{
  use base graphics to plot an SIR network
}
\author{
  Barry S Rowlingson
}

