\name{resetG}
\alias{resetG}
\title{resetG}
\usage{
  resetG(g)
}
\arguments{
  \item{g}{a graph network for SIR modelling}
}
\value{
  modified g
}
\description{
  reset graph
}
\details{
  reset all nodes to "S", reset clock to start, clear
  infection/recovery times
}
\author{
  Barry S Rowlingson
}

