\name{stopWhenClear}
\alias{stopWhenClear}
\title{stopWhenClear}
\usage{
stopWhenClear(g)
}
\arguments{
  \item{g}{an SIR graph}
}
\value{
TRUE if ther are no infectious cases, otherwise FALSE
}
\description{
Stop when no more infectious cases
}
\details{
This is a stopper function that returns true if there are
no infectious cases.
}
\author{
Barry S Rowlingson
}

