\name{timePlot}
\alias{timePlot}
\title{timePlot}
\usage{
  timePlot(g, t, n, s)
}
\arguments{
  \item{g}{an SIR graph}

  \item{t}{time points for plotting or...}

  \item{n}{number of time points between start and finish
  or...}

  \item{s}{step size between start and finish}
}
\value{
  a ggplot graph object
}
\description{
  plot SIR by time
}
\details{
  produces a line plot of S I R numbers
}
\author{
  Barry S Rowlingson
}

